Back to Visual Code

+++ Mastering_Critical_Skills/Skills/Section_04_Data_Types_and_Variables/Section_04_HW_01_Math_Operations.cpp
@@ -4,8 +4,8 @@
 #include <iostream>
 
 auto main() -> int {
-    int a {};
-    int b {};
+    double a {};
+    double b {};
     std::cout << "Enter two numbers: ";
     std::cin >> a >> b;
